% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerEdge.R, R/ConquerHive.R,
%   R/ConquerLocusZoom.R, R/ConquerRing.R, R/ConquerViolin.R
\name{conquer.d3js-shiny}
\alias{conquer.d3js-shiny}
\alias{ConquerEdgeOutput}
\alias{renderConquerEdge}
\alias{ConquerHiveOutput}
\alias{renderConquerHive}
\alias{ConquerLocusZoomOutput}
\alias{renderConquerLocusZoom}
\alias{ConquerRingOutput}
\alias{renderConquerRing}
\alias{ConquerViolinOutput}
\alias{renderConquerViolin}
\title{Shiny bindings for conquer.d3js}
\usage{
ConquerEdgeOutput(outputId, width = "100\%", height = "400px")

renderConquerEdge(expr, env = parent.frame(), quoted = FALSE)

ConquerHiveOutput(outputId, width = "100\%", height = "400px")

renderConquerHive(expr, env = parent.frame(), quoted = FALSE)

ConquerLocusZoomOutput(outputId, width = "100\%", height = "400px")

renderConquerLocusZoom(expr, env = parent.frame(), quoted = FALSE)

ConquerRingOutput(outputId, width = "100\%", height = "400px")

renderConquerRing(expr, env = parent.frame(), quoted = FALSE)

ConquerViolinOutput(outputId, width = "100\%", height = "400px")

renderConquerViolin(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a conquer.d3js}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using conquer.d3js within Shiny
applications and interactive Rmd documents.

Output and render functions for using conquer.d3js within Shiny
applications and interactive Rmd documents.

Output and render functions for using conquer.d3js within Shiny
applications and interactive Rmd documents.

Output and render functions for using conquer.d3js within Shiny
applications and interactive Rmd documents.

Output and render functions for using conquer.d3js within Shiny
applications and interactive Rmd documents.
}
