% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerEdge.R
\name{ConquerEdge}
\alias{ConquerEdge}
\title{ConquerEdge}
\usage{
ConquerEdge(conquerSummary, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{conquerSummary}{Summary file  of CONQUER}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{elementId}{elementId}
}
\description{
ConquerEdge
}
