% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerHive.R
\name{ConquerHive}
\alias{ConquerHive}
\title{ConquerHive}
\usage{
ConquerHive(
  SNPData,
  cis = T,
  trans = T,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{SNPData}{SNPdata}

\item{cis}{Include cis}

\item{trans}{Include trans}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{elementId}
}
\description{
ConquerHive
}
