% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerLocusZoom.R
\name{ConquerLocusZoom}
\alias{ConquerLocusZoom}
\title{ConquerLocusZoom}
\usage{
ConquerLocusZoom(SNPData, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{SNPData}{SNPdata}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{elementId}
}
\description{
ConquerLocusZoom
}
