% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerRing.R
\name{ConquerRing}
\alias{ConquerRing}
\title{ConquerRing}
\usage{
ConquerRing(
  conquerSummary,
  tissue,
  KEGG_DATA,
  hoverID,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
ConquerRing
}
