% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerViolin.R
\name{ConquerViolin}
\alias{ConquerViolin}
\title{ConquerViolin}
\usage{
ConquerViolin(SNP, gene, tissue, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{SNP}{SNP name}

\item{gene}{gene name}

\item{tissue}{tissue}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{elementId}
}
\description{
ConquerViolin
}
