% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_CLASSIFICATION.R
\docType{data}
\name{KEGG_CLASSIFICATION}
\alias{KEGG_CLASSIFICATION}
\title{A dataframe with KEGG classification}
\format{
A [[dataframe]] object
}
\usage{
KEGG_CLASSIFICATION
}
\description{
A dataframe with KEGG classification
}
\keyword{datasets}
