% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConquerRing.R
\name{ConquerRing}
\alias{ConquerRing}
\title{ConquerRing}
\usage{
ConquerRing(
  conquerSummary,
  tissue,
  KEGG_DATA,
  hoverID,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{conquerSummary}{conquerSummary}

\item{tissue}{Tissue of interest}

\item{KEGG_DATA}{KEGG_DATA generated by CONQUER}

\item{hoverID}{hoverID}

\item{width}{Width}

\item{height}{Height}

\item{elementId}{elementId}
}
\description{
ConquerRing
}
